+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - (Maintenance) Generate CRD with Schemas
 - (Feature) DebugPackage Improvements
 - (Feature) Improve Bootstrap time
+- (Feature) Move Condition Actions to Internal
 
 ## [1.2.47](https://github.com/arangodb/kube-arangodb/tree/1.2.47) (2025-03-28)
 - (Bugfix) Use Profile Annotations